#17 Remove obsolete sky model

+++ src/main/java/igentuman/galacticresearch/sky/body/ISkyBody.java
@@ -11,7 +11,7 @@ public interface ISkyBody {
 
     boolean isVisible();
 
-    int getSize();
+    float getSize();
 
     default long clientTime()
     {
